use redux to manage colorScheme
