+++ packages/core-mobile/app/new/common/components/BlurViewWithFallback.tsx
@@ -2,7 +2,8 @@ import { BlurView } from 'expo-blur'
 import React, { useMemo } from 'react'
 import { ViewStyle, Platform } from 'react-native'
 import { alpha, useTheme, View } from '@avalabs/k2-alpine'
-import { useColorScheme } from 'common/contexts/ColorSchemeProvider'
+import { useSelector } from 'react-redux'
+import { selectSelectedColorScheme } from 'store/settings/appearance'
 
 export const BlurViewWithFallback = ({
   children,
@@ -12,7 +13,7 @@ export const BlurViewWithFallback = ({
   style?: ViewStyle
 }): JSX.Element => {
   const { theme } = useTheme()
-  const { colorScheme } = useColorScheme()
+  const colorScheme = useSelector(selectSelectedColorScheme)
 
   const iosContainerStyle = useMemo(
     () => [
