+++ packages/core-mobile/app/new/common/hooks/useLogoModal.tsx
@@ -1,14 +1,15 @@
 import { K2AlpineThemeProvider } from '@avalabs/k2-alpine'
 import { LogoModal } from 'common/components/LogoModal'
-import { useColorScheme } from 'common/contexts/ColorSchemeProvider'
 import React from 'react'
+import { useSelector } from 'react-redux'
+import { selectSelectedColorScheme } from 'store/settings/appearance'
 import { hideModal, showModal } from 'utils/modal'
 
 export const useLogoModal = (): {
   showLogoModal: () => void
   hideLogoModal: () => void
 } => {
-  const { colorScheme } = useColorScheme()
+  const colorScheme = useSelector(selectSelectedColorScheme)
 
   const showLogoModal = (): void => {
     showModal(
