+++ packages/core-mobile/app/store/settings/appearance/slice.ts
@@ -1,5 +1,6 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 import { RootState } from 'store'
+import { ColorSchemeName } from 'react-native'
 import { Appearance, initialState } from './types'
 
 const reducerName = 'appearance'
@@ -10,6 +11,9 @@ export const appearanceSlice = createSlice({
   reducers: {
     setSelectedAppearance: (state, action: PayloadAction<Appearance>) => {
       state.selected = action.payload
+    },
+    setSelectedColorScheme: (state, action: PayloadAction<ColorSchemeName>) => {
+      state.colorScheme = action.payload
     }
   }
 })
@@ -19,7 +23,14 @@ export const selectSelectedAppearance = (state: RootState): Appearance => {
   return state.settings.appearance.selected
 }
 
+export const selectSelectedColorScheme = (
+  state: RootState
+): ColorSchemeName => {
+  return state.settings.appearance.colorScheme
+}
+
 // actions
-export const { setSelectedAppearance } = appearanceSlice.actions
+export const { setSelectedAppearance, setSelectedColorScheme } =
+  appearanceSlice.actions
 
 export const appearanceReducer = appearanceSlice.reducer
